Move classes into separate files and some methods into separate classes
